Initial Upload
Added the following class files:

ContourDrawer
Line
PlayerCamera

ContourDrawer is the main class for generating contour lines from an image. It takes in an image and uses a marching squares algorithm to generate lines at intervals of the isovalue. Isovalue starts at 255 and each loop is reduced by contourInterval (how many colour values to jump by for each pass, i.e. how many values of that 255 = one contour interval, noting that standard contour lines are at intervals of 10m). It then uses linear interpolation to smooth out the points, and each point is stored as a Line class before a function iterates over each Line to find the one in front of it to make a Line2D.

Line is a simple class which contains two Vector2s - one for the start point and one for the end.

PlayerCamera is a basic camera controller - press in middle mouse button to move camera around. Zoom in and out with mouse wheel.