Smoothing and spot reduction
Added quadratic and cubic bezier curve function for ContourLine node. This smooths out all lines to appear less jagged.

Added function for getting area of each line and removing any lines with area smaller than a set value (this is to remove artefact lines which are very small or are straight lines etc).

TODO:
There is a better way to calculate the area but for now it just adds the squared distance of each point to the root point and divides the result by the number of total points.